+++ cvat-core/src/enums.js
@@ -403,6 +403,24 @@
         OWNER: 'owner',
     });
 
+    /**
+     * Sorting methods
+     * @enum {string}
+     * @name SortingMethod
+     * @memberof module:API.cvat.enums
+     * @property {string} LEXICOGRAPHICAL 'lexicographical'
+     * @property {string} NATURAL 'natural'
+     * @property {string} PREDEFINED 'predefined'
+     * @property {string} RANDOM 'random'
+     * @readonly
+     */
+    const SortingMethod = Object.freeze({
+        LEXICOGRAPHICAL: 'lexicographical',
+        NATURAL: 'natural',
+        PREDEFINED: 'predefined',
+        RANDOM: 'random',
+    });
+
     module.exports = {
         ShareFileType,
         TaskStatus,
@@ -422,5 +440,6 @@
         CloudStorageProviderType,
         CloudStorageCredentialsType,
         MembershipRole,
+        SortingMethod,
     };
 })();