+++ cvat-core/src/session.js
@@ -1112,6 +1112,7 @@
                 copy_data: undefined,
                 dimension: undefined,
                 cloud_storage_id: undefined,
+                sorting_method: undefined,
             };
 
             const updateTrigger = new FieldUpdateTrigger();
@@ -1647,6 +1648,16 @@
                     cloudStorageId: {
                         get: () => data.cloud_storage_id,
                     },
+                    sortingMethod: {
+                        /**
+                         * @name sortingMethod
+                         * @type {module:API.cvat.enums.SortingMethod}
+                         * @memberof module:API.cvat.classes.Task
+                         * @instance
+                         * @readonly
+                         */
+                        get: () => data.sorting_method,
+                    },
                     _internalData: {
                         get: () => data,
                     },
@@ -2138,6 +2149,7 @@
             image_quality: this.imageQuality,
             use_zip_chunks: this.useZipChunks,
             use_cache: this.useCache,
+            sorting_method: this.sortingMethod,
         };
 
         if (typeof this.startFrame !== 'undefined') {