+++ cvat/apps/engine/backup.py
@@ -65,6 +65,7 @@ class _TaskBackupBase():
             'chunk_type',
             'storage_method',
             'storage',
+            'sorting_method',
         }
 
         self._prepare_meta(allowed_fields, data)