+++ cvat/apps/engine/views.py
@@ -237,12 +237,17 @@ class ProjectFilter(filters.FilterSet):
 @method_decorator(name='retrieve', decorator=swagger_auto_schema(operation_summary='Method returns details of a specific project'))
 @method_decorator(name='destroy', decorator=swagger_auto_schema(operation_summary='Method deletes a specific project'))
 @method_decorator(name='partial_update', decorator=swagger_auto_schema(operation_summary='Methods does a partial update of chosen fields in a project'))
+
 class ProjectViewSet(viewsets.ModelViewSet):
-    queryset = models.Project.objects.all().order_by('-id')
+    queryset = models.Project.objects.prefetch_related(Prefetch('label_set',
+        queryset=models.Label.objects.order_by('id')
+    ))
+
     search_fields = ("name", "owner__username", "assignee__username", "status")
     filterset_class = ProjectFilter
     ordering_fields = ("id", "name", "owner", "status", "assignee")
-    http_method_names = ['get', 'post', 'head', 'patch', 'delete']
+    ordering = ("-id",)
+    http_method_names = ('get', 'post', 'head', 'patch', 'delete')
 
     def get_serializer_class(self):
         if self.request.path.endswith('tasks'):
@@ -472,7 +477,8 @@ class DjangoFilterInspector(CoreAPICompatInspector):
 @method_decorator(name='destroy', decorator=swagger_auto_schema(operation_summary='Method deletes a specific task, all attached jobs, annotations, and data'))
 @method_decorator(name='partial_update', decorator=swagger_auto_schema(operation_summary='Methods does a partial update of chosen fields in a task'))
 class TaskViewSet(viewsets.ModelViewSet):
-    queryset = Task.objects.all().prefetch_related(
+    queryset = Task.objects.prefetch_related(
+            Prefetch('label_set', queryset=models.Label.objects.order_by('id')),
             "label_set__attributespec_set",
             "segment_set__job_set",
         ).order_by('-id')
@@ -672,9 +678,8 @@ class TaskViewSet(viewsets.ModelViewSet):
             db_task.data = db_data
             db_task.save()
             data = {k:v for k, v in serializer.data.items()}
-            data['use_zip_chunks'] = serializer.validated_data['use_zip_chunks']
-            data['use_cache'] = serializer.validated_data['use_cache']
-            data['copy_data'] = serializer.validated_data['copy_data']
+            for extra_key in { 'use_zip_chunks', 'use_cache', 'copy_data' }:
+                data[extra_key] = serializer.validated_data[extra_key]
             if data['use_cache']:
                 db_task.data.storage_method = StorageMethodChoice.CACHE
                 db_task.data.save(update_fields=['storage_method'])